+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.7.0
 
 * Spec includes `deps` so you can just pass one object.
+* Snapshot data includes `deps`. Removed `dag`.
 * Get error stacktrace if available.
 * Wrap initial input data in an array to match other nodes.